+++ wox.ui.flutter/wox/lib/entity/wox_setting.dart
@@ -39,6 +39,7 @@ class WoxSetting {
   late String httpProxyUrl;
   late bool enableAutoBackup;
   late bool enableAutoUpdate;
+  late String releaseChannel;
   late bool enableAnonymousUsageStats;
   late String customPythonPath;
   late String customNodejsPath;
@@ -82,6 +83,7 @@ class WoxSetting {
     required this.httpProxyUrl,
     required this.enableAutoBackup,
     required this.enableAutoUpdate,
+    required this.releaseChannel,
     required this.enableAnonymousUsageStats,
     required this.customPythonPath,
     required this.customNodejsPath,
@@ -166,6 +168,7 @@ class WoxSetting {
     httpProxyUrl = json['HttpProxyUrl'] ?? '';
     enableAutoBackup = json['EnableAutoBackup'] ?? false;
     enableAutoUpdate = json['EnableAutoUpdate'] ?? true;
+    releaseChannel = json['ReleaseChannel'] ?? 'stable';
     enableAnonymousUsageStats = json['EnableAnonymousUsageStats'] ?? true;
     customPythonPath = json['CustomPythonPath'] ?? '';
     customNodejsPath = json['CustomNodejsPath'] ?? '';
@@ -211,6 +214,7 @@ class WoxSetting {
     data['HttpProxyUrl'] = httpProxyUrl;
     data['EnableAutoBackup'] = enableAutoBackup;
     data['EnableAutoUpdate'] = enableAutoUpdate;
+    data['ReleaseChannel'] = releaseChannel;
     data['EnableAnonymousUsageStats'] = enableAnonymousUsageStats;
     data['CustomPythonPath'] = customPythonPath;
     data['CustomNodejsPath'] = customNodejsPath;